+++ packages/browser-wallet/src/popup/pages/Setup/RecoveryPhrase.tsx
@@ -43,10 +43,10 @@ export function EnterRecoveryPhrase() {
 
     return (
         <>
-            <PageHeader canGoBack>Your recovery phrase</PageHeader>
+            <PageHeader canGoBack>{t('recoveryPhrase.title')}</PageHeader>
             <div className="onboarding-setup__page-with-header">
                 <div className="onboarding-setup__page-with-header__description">
-                    {t('enterRecoveryPhrase.description')}
+                    {t('confirmRecoveryPhrase.description')}
                 </div>
                 <div className="p-10">
                     <Form<FormValues> onSubmit={handleSubmit}>