+++ packages/browser-wallet/src/popup/store/utils.ts
@@ -14,6 +14,7 @@ import {
     storedSelectedIdentity,
     storedSeedPhrase,
     storedIdentityProviders,
+    sessionAccountInfoCache,
 } from '@shared/storage/access';
 import { ChromeStorageKey } from '@shared/storage/types';
 import { atom, WritableAtom } from 'jotai';
@@ -32,6 +33,7 @@ const accessorMap = {
     [ChromeStorageKey.SeedPhrase]: storedSeedPhrase,
     [ChromeStorageKey.IdentityProviders]: useIndexedStorage(storedIdentityProviders, getGenesisHash),
     [ChromeStorageKey.Passcode]: sessionPasscode,
+    [ChromeStorageKey.AccountInfoCache]: useIndexedStorage(sessionAccountInfoCache, getGenesisHash),
 };
 
 export type AsyncWrapper<V> = {