+++ packages/browser-wallet/src/shared/storage/types.ts
@@ -1,4 +1,4 @@
-import type { IdentityObjectV1 } from '@concordium/web-sdk';
+import type { IdentityObjectV1, Versioned } from '@concordium/web-sdk';
 
 export enum ChromeStorageKey {
     ConnectedSites = 'connectedSites',
@@ -12,16 +12,12 @@ export enum ChromeStorageKey {
     Identities = 'identities',
     SelectedIdentity = 'selectedIdentity',
     IdentityProviders = 'identityProviders',
+    AccountInfoCache = 'accountInfoCache',
 }
 
 export enum Theme {
     Light = 'light',
     Dark = 'dark',
-}
-
-export enum Network {
-    Testnet = 1,
-    Mainnet = 919,
 }
 
 /**
@@ -37,7 +33,7 @@ export interface BaseIdentity {
     status: CreationStatus;
     name: string;
     index: number;
-    provider: number;
+    providerIndex: number;
 }
 
 export interface PendingIdentity extends BaseIdentity {
@@ -52,10 +48,7 @@ export interface RejectedIdentity extends BaseIdentity {
 
 export interface ConfirmedIdentity extends BaseIdentity {
     status: CreationStatus.Confirmed;
-    idObject: {
-        v: 0;
-        value: IdentityObjectV1;
-    };
+    idObject: Versioned<IdentityObjectV1>;
 }
 
 export type Identity = PendingIdentity | RejectedIdentity | ConfirmedIdentity;
@@ -87,6 +80,7 @@ export interface IpInfo {
  */
 export interface IdentityProviderMetaData {
     issuanceStart: string;
+    recoveryStart: string;
     icon: string;
     support: string;
 }
@@ -121,6 +115,7 @@ export interface BaseCredential {
     credNumber: number;
     status: CreationStatus;
     identityIndex: number;
+    providerIndex: number;
     // Policy (is in accountInfo)
     // CredentialIndex = 0
 }