+++ packages/browser-wallet/CHANGELOG.md
@@ -6,6 +6,13 @@
 
 -   Toasts for showing errors when failing to retrieve account balance or transaction history.
 -   Send CCD page.
+-   Transaction history is now updated on balance changes, so that the new transactions are retrieved.
+-   Testnet faucet for getting some CCD for testing.
+-   Terms and conditions to setup flow and the about page.
+
+### Changed
+
+-   Account information retrieved from the JSON-RPC server is now cached for better performance, and a more responsive UI.
 
 ### Fixed